NLP-18 Separate incident type steps into their own preprocessors
